refactored code: desktop/mobile wrapper, created dropdown component for advertisers page, fixed scss, changed modal text ternaries
